+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 - Pass `options` to `key` pred.
 - Pass `getFrom` instead of `goTo` to `_validate` in `or`.
+- Always return promise on valid result.
 
 ### Deprecate