+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Pass `options` to `key` pred.
 - Pass `getFrom` instead of `goTo` to `_validate` in `or`.
 - Always return promise on valid result.
+- Ensure value is always attached to `validate` result.
 
 ### Deprecate