+++ CHANGELOG.md
@@ -6,6 +6,18 @@
 
 ### Grow
 
+### Fix
+
+### Deprecate
+
+---
+
+## [2.2.1] - 2020-11-16
+
+### Break
+
+### Grow
+
 - Set arity on combined pred function based on argument pred functions. Allows checking arity before creating/passing `getFrom` function.
 - Expose `validatePred` and `getPath` utility function.